Removing flag android.service.notification.notification_silent_flag from RELEASE_CONFIG_TRUNKFOOD_STAGING.
This commit was automatically created by advancing a workflow in Gantry. Manual edits to this file are forbidden. If you need to rollback, please visit  to begin a rollback.

POSTSUBMIT ERRORS: If this change breaks postsubmits, please revert it and attach the associated bugs to the reverted CL's description.

Change-Id: I312361105169d9d38963303cd22221dae0208701
Flag: android.service.notification.notification_silent_flag
Ignore-AOSP-First: AConfig change for future AOSP release