Removing flag android.service.notification.notification_force_grouping from RELEASE_CONFIG_TRUNKFOOD_STAGING.
This commit was automatically created by advancing a workflow in Gantry. Manual edits to this file are forbidden. If you need to rollback, please visit  to begin a rollback.

POSTSUBMIT ERRORS: If this change breaks postsubmits, please revert it and attach the associated bugs to the reverted CL's description.

Change-Id: Id66a8a1acc686f1c5e24e20785fe9aaa05de4be0
Flag: android.service.notification.notification_force_grouping
Ignore-AOSP-First: AConfig change for future AOSP release