Removing flag com.android.systemui.media_controls_refactor from RELEASE_CONFIG_TRUNKFOOD_STAGING.
This commit was automatically created by advancing a workflow in Gantry. Manual edits to this file are forbidden. If you need to rollback, please visit  to begin a rollback.

POSTSUBMIT ERRORS: If this change breaks postsubmits, please revert it and attach the associated bugs to the reverted CL's description.

Change-Id: I28dedabe6efd0d76d786f6d8ec2d88256dbc55ff
Flag: com.android.systemui.media_controls_refactor
Ignore-AOSP-First: AConfig change for future AOSP release