Removing flag com.android.providers.settings.support_overrides from RELEASE_CONFIG_TRUNKFOOD_STAGING.
This commit was automatically created by advancing a workflow in Gantry. Manual edits to this file are forbidden. If you need to rollback, please visit  to begin a rollback.

POSTSUBMIT ERRORS: If this change breaks postsubmits, please revert it and attach the associated bugs to the reverted CL's description.

Change-Id: I2c8a77f2a5adbb7e57e394eca0aed3d4dcf85cc9
Flag: com.android.providers.settings.support_overrides
Ignore-AOSP-First: AConfig change for future AOSP release