Removing flag android.server.allow_removing_vpn_service from RELEASE_CONFIG_TRUNKFOOD_STAGING.
This commit was automatically created by advancing a workflow in Gantry. Manual edits to this file are forbidden. If you need to rollback, please visit  to begin a rollback.

POSTSUBMIT ERRORS: If this change breaks postsubmits, please revert it and attach the associated bugs to the reverted CL's description.

Flag: android.server.allow_removing_vpn_service
Ignore-AOSP-First: AConfig change for future AOSP release

Bug: 340928692
Change-Id: I874d7b01670b3279855628dda609273162f41138